+++ examples/python/ml/torch_lr_experiment/torch_lr_experiment.py
@@ -22,18 +22,12 @@ from sklearn import metrics
 
 import spu.utils.distributed as ppd
 
-# This is an experimental example to show legacy pytorch program could be run
-# by SPU. Currently we rely on torch-mlir to convert torch code into MLIR
-# (specifically MHLO) which is then consumed by SPU. To run this example,
-# torch-mlir python package should be installed. This example here trains a
-# linear regression model in plaintext and makes private inferences with joint
-# features.
 
 # Start nodes.
 # > bazel run -c opt //examples/python/utils:nodectl -- up
 #
 # Run this example script.
-# > bazel run -c opt //examples/python/ml/torch_experiment:torch_experiment
+# > bazel run -c opt //examples/python/ml/torch_lr_experiment:torch_lr_experiment
 
 
 class LinearRegression(torch.nn.Module):
@@ -41,17 +35,15 @@ class LinearRegression(torch.nn.Module):
         super(LinearRegression, self).__init__()
         self.linear = torch.nn.Linear(30, 1)
 
-    def forward(self, x1, x2):
-        y_pred = self.linear(torch.cat((x1, x2), 1))
+    def forward(self, x):
+        y_pred = self.linear(x)
         return y_pred
 
 
 def train(model, n_epochs=500, lr=0.01):
     print('Train model with plaintext features\n------\n')
     x, y = breast_cancer()
-    x1, x2 = x[:, :15], x[:, 15:]
-    x1 = torch.Tensor(x1)
-    x2 = torch.Tensor(x2)
+    x = torch.Tensor(x)
     y = torch.Tensor(y).view(-1, 1)
 
     criterion = torch.nn.BCEWithLogitsLoss()
@@ -59,7 +51,7 @@ def train(model, n_epochs=500, lr=0.01):
     optimizer = torch.optim.SGD(model.parameters(), lr=lr)
 
     for _ in range(n_epochs):
-        pred_y = model(x1, x2)
+        pred_y = model(x)
         loss = criterion(pred_y, y)
         optimizer.zero_grad()
         loss.backward()
@@ -70,7 +62,6 @@ def train(model, n_epochs=500, lr=0.01):
 
 # prepare test datasets
 def breast_cancer(
-    col_slicer=slice(None, None, None),
     train: bool = True,
     *,
     normalize: bool = True,
@@ -96,7 +87,6 @@ def breast_cancer(
     else:
         x_ = x_test
         y_ = y_test
-    x_ = x_[:, col_slicer]
     return x_.astype(dtype=np.float32), y_.astype(dtype=np.float32)
 
 
@@ -105,10 +95,10 @@ import time
 
 def run_inference_on_cpu(model):
     print('Run on CPU\n------\n')
-    x_test, y_test = breast_cancer(slice(None, None, None), False)
-    x1, x2 = torch.Tensor(x_test[:, :15]), torch.Tensor(x_test[:, 15:])
+    x_test, y_test = breast_cancer(False)
+    x = torch.Tensor(x_test)
     start_ts = time.time()
-    y_pred = model.forward(x1, x2).cpu().detach().numpy()
+    y_pred = model(x).cpu().detach().numpy()
     end_ts = time.time()
     auc = metrics.roc_auc_score(y_test, y_pred)
     print(f"AUC(cpu)={auc}, time={end_ts-start_ts}\n------\n")
@@ -123,33 +113,34 @@ with open(args.config, 'r') as file:
 
 ppd.init(conf["nodes"], conf["devices"], framework=ppd.Framework.EXP_TORCH)
 
+from collections import OrderedDict
+from jax.tree_util import tree_map
+
 
 def run_inference_on_spu(model):
     print('Run on SPU\n------\n')
-    x1, _ = ppd.device("P1")(breast_cancer)(slice(None, 15), False)
-    x2, _ = ppd.device("P2")(breast_cancer)(slice(15, None), False)
+
+    # load parameters and buffers on P1
+    params_buffers = OrderedDict()
+    for k, v in model.named_parameters():
+        params_buffers[k] = v
+    for k, v in model.named_buffers():
+        params_buffers[k] = v
+    params = ppd.device("P1")(
+        lambda input: tree_map(lambda x: x.detach().numpy(), input)
+    )(params_buffers)
+
+    # load inputs on P2
+    x, _ = ppd.device("P2")(breast_cancer)(False)
+
     start_ts = time.time()
-    y_pred_ciphertext = ppd.device('SPU')(model)(x1, x2)
+    y_pred_ciphertext = ppd.device('SPU')(model)(params, x)
     end_ts = time.time()
     y_pred_plaintext = ppd.get(y_pred_ciphertext)
-    _, y_test = breast_cancer(slice(None, None, None), False)
+    _, y_test = breast_cancer(False)
     auc = metrics.roc_auc_score(y_test, y_pred_plaintext)
     print(f"AUC(cpu)={auc}, time={end_ts-start_ts}\n------\n")
     return auc
-
-
-def compile_torch_to_mhlo(model):
-    print('Compile torch program to mhlo test\n------\n')
-    x_test, _ = breast_cancer(slice(None, None, None), False)
-    x1, x2 = torch.Tensor(x_test[:, :15]), torch.Tensor(x_test[:, 15:])
-    import torch_mlir
-
-    module = torch_mlir.compile(
-        model,
-        [x1, x2],
-        output_type=torch_mlir.OutputType.MHLO,
-    )
-    print(f"MHLO={module}\n------\n")
 
 
 if __name__ == '__main__':
@@ -159,8 +150,7 @@ if __name__ == '__main__':
     model = LinearRegression()
     # Train model with plaintext features
     train(model)
-    # Torch-mlho conversion test
-    compile_torch_to_mhlo(model)
+    model.eval()
     # Native torch inference
     run_inference_on_cpu(model)
     # SPU inference