+++ libspu/compiler/core/core.cc
@@ -62,8 +62,6 @@ void Core::buildPipeline(mlir::PassManager *pm) {
     optPM.addPass(mlir::pphlo::createRewriteDivSqrtPatterns());
   }
 
-  optPM.addPass(mlir::pphlo::createExpandSecretGatherPass());
-
   if (options.enable_optimize_denominator_with_broadcast()) {
     optPM.addPass(mlir::pphlo::createOptimizeDenominatorWithBroadcast());
   }